Lesson13: Implementing RecyclerView: fixed problem with groupie library. Need to be fixed: displaying weather icon in FutureWeatherFragment

+++ app/src/main/java/com/learning/forecastweathermmvmapp/data/db/entity/Day.kt
@@ -11,9 +11,9 @@ data class Day(
     @SerializedName("avgtemp_f")
     val avgtempF: Double,
     @SerializedName("avgvis_km")
-    val avgvisKm: Int,
+    val avgvisKm: Double,
     @SerializedName("avgvis_miles")
-    val avgvisMiles: Int,
+    val avgvisMiles: Double,
     @Embedded(prefix = "condition_")
     val condition: Condition,
     @SerializedName("maxtemp_c")
@@ -21,7 +21,7 @@ data class Day(
     @SerializedName("maxtemp_f")
     val maxtempF: Double,
     @SerializedName("maxwind_kph")
-    val maxwindKph: Int,
+    val maxwindKph: Double,
     @SerializedName("maxwind_mph")
     val maxwindMph: Double,
     @SerializedName("mintemp_c")
@@ -29,8 +29,8 @@ data class Day(
     @SerializedName("mintemp_f")
     val mintempF: Double,
     @SerializedName("totalprecip_in")
-    val totalprecipIn: Int,
+    val totalprecipIn: Double,
     @SerializedName("totalprecip_mm")
-    val totalprecipMm: Int,
-    val uv: Int
+    val totalprecipMm: Double,
+    val uv: Double
 )
